// src/utils/enum.ts
export enum GENDERS {
  MALE = "male",
  FEMALE = "female",
}

export enum GOALS {
  LOSE = "lose",
  MAINTAIN = "maintain",
  GAIN = "gain",
}

export enum ACTIVITY_LEVELS {
  SEDENTARY = "sedentary",
  LIGHTLY_ACTIVE = "lightly active",
  MODERATELY_ACTIVE = "moderately active",
  VERY_ACTIVE = "very active",
  SUPER_ACTIVE = "super active",
}